Make keyframe updating automatic
This also refactors a lot of other things that I needed to change in order to make this work.
Added Parent and Children properties to GameObject
AddedRotation and Scale to Object2D
Made Enemy radius calculated from the Scale property
Added AnimationPlayer which updates its parent's position, rotation and scale.